Create a new migration to fix special characters in the rabbitmq password
Previously the arguments to the find_by call were incorrect so
the migration would never find a record. The issue is that the
userid was "tower" not "ansible".

In the new version we only search by the type and authtype attributes
which should never change and are clearly defined in the application code

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1678337
